refactor: CodeCheck, VerifyModel 컴포넌트화
