+++ src/components/auth/register/verify/CodeCheck/style.ts
@@ -1,0 +1,22 @@
+import { motion } from 'framer-motion';
+import styled from '@emotion/styled';
+import Button from '@/components/common/Button';
+
+export const CodeDiv = styled(motion.div)`
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+`;
+
+export const ReSendButton = styled(Button)`
+  align-self: flex-end;
+  padding: 0;
+  font-size: 0.8rem;
+  font-weight: 600;
+  color: #94a3b8;
+  transition: all 0.1s;
+
+  &:hover {
+    color: #5d5d5d;
+  }
+`;
